CloudFunctions/insertbd/src/requirements.txt
google-cloud-storage
google-cloud-secret-manager
mysql-connector-python
xmltodict
numpy